Changes to make the project run under Windows (Visual Studio Community and MSVC compiler)

+++ src/audio/ail/bank.h
@@ -355,8 +355,8 @@ public:
      */
     static MidiBank emptyBank(uint16_t index = 0);
 
-    inline int countMelodic() const { return Ins_Melodic_box.size(); }
-    inline int countDrums() const   { return Ins_Percussion_box.size(); }
+    inline int countMelodic() const { return (int)(Ins_Melodic_box.size()); }
+    inline int countDrums() const   { return (int)(Ins_Percussion_box.size()); }
 
     /**
      * @brief Get the identified bank
